gcc7: Fix i86pc build.
Attempting to build this using the pkgsrc gcc7 package results in a sort
of cross-compile from x86_64 to i386, which results in header problems
when building libgcc.  Force the host to the same as the target to work
around this.

Also use the just-built binutils for tools, avoids an issue seen where
i386-*-nm couldn't be found.